requirements.txt
streamlit
llama-index
nest_asyncio
python-dotenv
llama-parse
